Add the checkmark icon from icomoon. It is used as a button to confirm a tagged face.
Add script to generate the release ZIP.

Convert capture.php into a class to fix issues with the Logger.

Redesign the client / server interface implemented in admin_mug.js and capture.php
	It is now possible to have faces defined without an identifying tag.

	Faces can be hidden, it is as if they were removed but the entry is still in the database. This prevents facial detection software from reinserting an unwanted or bogus entry in the case that facial detection is rerun on the same photo.

Updated the database table face_tag_positions to support automatic face detection.
	New primary key column "id" so that tag_id can be null.

	Added new column "face_index" to match the row to the results from the facial detection code.

	tag_id can now be null. This is for the case where a face has been detected but not identified.

	Added new column "confirmed" which is initially false when a new tag is added by the facial matching component. It is automatically set to true in any case where the user is associating a tag.

	Added new column "ignored" which is set when a row created by facial detection is "removed".

Added a new menu block on the left side of the window called MugShot. It currently has two items "Unidentified People" and "Unconfirmed Matched People".
	Unidentified People are those that facial detection has found but which haven't been assigned a face tag. The tag_id is set to null.

	Unconfirmed Matched People are those that the software has matched to another face but which hasn't been reviewed and confirmed by a human.

All facial detection and matching has been moved to a standalone app. In fact it has been specifically designed to run on a different more powerful machine that can be managed by the installer while using an inexpensive web service to run the Gallery.

Confirmed faces are indicated by the bounding box being drawn in green.

When a double click is done within the bounding box the name box is displayed and focus is transfered to it.